+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 ## CHANGELOG
 
-1. refactored theme switch code.
+1. StackView is its own component.
+2. `shift + p` to change subreddit.
+3. `s` to switch to FocusView and start with first post.
+4. changed upVote icon to BiUpvote from react-icons.